feat(storage): widen persistent coordinate storage to double (REQ-101 ±0.002 ft, ADR-054 Phase A, #440)
TASK-228 Phase A. The four flat geometry stores and every entity scalar
coordinate field move from float to double so a stored coordinate is within
±0.002 ft at any drawing magnitude (REQ-101, D-2026-09-08-i). Strides and the
interleaved-XYZ layout are unchanged (architecture §11.8).

Stores widened in all three copies (AppCommandState, DrawingGeometrySnapshot,
DrawingTab):
- userLinesFlat, userPolylineVerts, userCirclesCxCyZR, featureLineVerts
- CadFilledRegion::vertsXyz
- CadArc (cx/cy/r/startRad/sweepRad/z), CadEllipse (cx/cy/majVx/majVy/ratio/z)
- CadClipboard and cadblock.hpp block-content stores (lines/circles/polyVerts)

The float narrowing now happens once, at the GPU-upload boundary (ADR-054 (b)):
RenderScene() and AppendChainEdgesVc() take const std::vector<double>&; the GL
vertex format is unchanged and the narrowing is implicit at buffer assembly.

Shared coordinate helpers are templated on scalar type so no call site changes
(RotateAroundBase, ReflectPtAcrossLine, ScalePtAroundBase, HelmertPt,
ArcFromAngles, {rotate,scale,mirror}PreviewPt, ChainHitsRect, PointInClosedPoly,
ClosedPolylineCentroid, PolylineLoopMeanZ, ClosedPolyGeometricPickDistSq,
CheckFinite, CadChainHasEntities, and the CadCoordinateFrame / paper-transform /
ScaleSelectionZAboutBase lambdas). UI Properties editors (PropGeomRow, paper
geomRow) route through ImGui::InputScalar with the matching ImGuiDataType.

IO: the .gs / DWG-trailer JSON reader now reads coordinate arrays as double, not
float — the writer already emitted full precision, so reading them back narrowed
broke round-trip idempotence (caught by regression-61 and fuzz-smoke). DXF and
DWG-native import still narrow at their static_cast<float> sites; those are
Phase B (#441). Paper-space stores (paper inches) stay float for now — small
magnitudes, not the ±0.002 ft case — also Phase B.

Build clean; ctest 1359/1359.

+++ src/render/ViewportRenderer.cpp
@@ -257,7 +257,7 @@ float LineweightMmToDevicePx(float mm) {
 /// any other 3D chain. The elevation-point flag is deliberately not consulted — an elevation point
 /// lies on the line, so including it changes nothing about the plan shape (ADR-035 (a)).
 void AppendChainEdgesVc(std::vector<float>& out, const CadExtendedGeometryInput& eg,
-                        const std::vector<float>* V, const std::vector<int>* O,
+                        const std::vector<double>* V, const std::vector<int>* O,
                         const std::vector<uint8_t>* Cl, const std::vector<EntityAttributes>* At,
                         float defR, float defG, float defB, float dashPatScale, double viewAnchorX,
                         double viewAnchorY) {
@@ -983,7 +983,7 @@ void ViewportRenderer::SetSize(int width, int height) {
 }
 
 void ViewportRenderer::RenderScene(const Camera& cam, int fbWidth, int fbHeight,
-                                   const std::vector<float>& userLines, const std::vector<float>& circlesCxCyZR,
+                                   const std::vector<double>& userLines, const std::vector<double>& circlesCxCyZR,
                                    std::uint32_t cadGpuRevision, const std::vector<float>& rubberLines,
                                    const CadSnap::Hit* snapOverlay, float snapGlyphHalfPx,
                                    const std::vector<float>* previewLines,
@@ -1987,7 +1987,7 @@ void ViewportRenderer::RenderScene(const Camera& cam, int fbWidth, int fbHeight,
         // on ITS OWN store. The feature-line append used to be nested inside the polyline block, so
         // it inherited whether polylines existed — a coupling with no reason behind it and a second
         // way for feature lines to vanish. REQ-087.
-        const auto appendChainStore = [&](const std::vector<float>* V, const std::vector<int>* O,
+        const auto appendChainStore = [&](const std::vector<double>* V, const std::vector<int>* O,
                                           const std::vector<uint8_t>* Cl,
                                           const std::vector<EntityAttributes>* At) {
           if (!CadChainHasEntities(V, O))
